Source data and cleaning script
I cannot get the github work for me with large files, so for now I uploaded the files it allows me to upload manually to give you reference. And the cleaned data is shared using my personal cloud:
http://gofile.me/6JGGa/Rk0LI3VPz
http://gofile.me/6JGGa/vTHLJrw3O
http://gofile.me/6JGGa/mBh1U1jBc
http://gofile.me/6JGGa/Frg9hab6L

Please let me know if you have issues downloading these files. 

Some information about how I cleaned the data:
For the first 3 correlations analysis, I only took the latest month(2021/03/31) from zillow data set since we don't need the time series. For the last one when we want to compare data between housing market and stock market, I saved the historical data. They both are saved to two files - one is for metro, and one is for suburb. It doesn't need the zipcode to metro mapping since the zillow file contains a column Metro, so if the metro is null, I think it means it is suburb. Let me know if you think differently. For the historical file, there are plenty nulls for certain months and certain zipcodes, and I'm not sure what the best way to handle these, so I will leave it for our group discussion. 

Please let me know anything else you want me to change or add to the data cleaning process to get you guys ready for the plot and analysis. We will have to talk about the github Tuesday to get it all working.